add useStore util for global stores
